Refactored the API services to use a common function to reduce code duplication.
